+++ lib/source/pl/core/ast/ast_node_bitfield_array_variable_decl.cpp
@@ -63,7 +63,7 @@ namespace pl::core::ast {
         std::vector<std::shared_ptr<ptrn::Pattern>> entries;
 
         size_t size = 0;
-        u128 entryIndex = 0;
+        u64 entryIndex = 0;
 
         auto addEntries = [&](std::vector<std::shared_ptr<ptrn::Pattern>> &&patterns) {
             for (auto &pattern : patterns) {
@@ -85,13 +85,13 @@ namespace pl::core::ast {
             err::E0001.throwError(fmt::format("Bitfield array was created with no size."), {}, this->getLocation());
 
         auto sizeNode = this->m_size->evaluate(evaluator);
-        std::variant<u128, ASTNodeWhileStatement *> boundsCondition;
+        std::variant<u64, ASTNodeWhileStatement *> boundsCondition;
 
         if (auto literalNode = dynamic_cast<ASTNodeLiteral *>(sizeNode.get()); literalNode != nullptr) {
             boundsCondition = std::visit(wolv::util::overloaded {
-                    [this](const std::string &) -> u128 { err::E0006.throwError("Cannot use string to index array.", "Try using an integral type instead.", this->getLocation()); },
-                    [this](const std::shared_ptr<ptrn::Pattern> &pattern) -> u128 {err::E0006.throwError(fmt::format("Cannot use custom type '{}' to index array.", pattern->getTypeName()), "Try using an integral type instead.", this->getLocation()); },
-                    [](auto &&size) -> u128 { return size; }
+                    [this](const std::string &) -> u64 { err::E0006.throwError("Cannot use string to index array.", "Try using an integral type instead.", this->getLocation()); },
+                    [this](const std::shared_ptr<ptrn::Pattern> &pattern) -> u64 {err::E0006.throwError(fmt::format("Cannot use custom type '{}' to index array.", pattern->getTypeName()), "Try using an integral type instead.", this->getLocation()); },
+                    [](auto &&size) -> u64 { return size; }
             }, literalNode->getValue());
         } else if (auto whileStatement = dynamic_cast<ASTNodeWhileStatement *>(sizeNode.get()); whileStatement != nullptr) {
             boundsCondition = whileStatement;
@@ -105,14 +105,14 @@ namespace pl::core::ast {
                 err::E0007.throwError(fmt::format("Bitfield array grew past set limit of {}", limit), "If this is intended, try increasing the limit using '#pragma array_limit <new_limit>'.", this->getLocation());
         };
 
-        if (std::holds_alternative<u128>(boundsCondition))
-            checkLimit(std::get<u128>(boundsCondition));
+        if (std::holds_alternative<u64>(boundsCondition))
+            checkLimit(std::get<u64>(boundsCondition));
 
-        u128 dataIndex = 0;
+        u64 dataIndex = 0;
 
         auto checkCondition = [&]() {
-            if (std::holds_alternative<u128>(boundsCondition))
-                return dataIndex < std::get<u128>(boundsCondition);
+            if (std::holds_alternative<u64>(boundsCondition))
+                return dataIndex < std::get<u64>(boundsCondition);
 
             checkLimit(entryIndex);
             return std::get<ASTNodeWhileStatement *>(boundsCondition)->evaluateCondition(evaluator);
